+++ src/api/objects/toot.ts
@@ -26,6 +26,7 @@ import {
     bracketed,
     collapseWhitespace,
     extractDomain,
+    htmlToParagraphs,
     htmlToText,
     isImage,
     isVideo,
@@ -65,6 +66,8 @@ const MIN_CHARS_FOR_LANG_DETECT = 8;
 const UNKNOWN = "unknown";
 const BLUESKY_BRIDGY = 'bsky.brid.gy';
 const REPAIR_TOOT = bracketed("repairToot");
+const HASHTAG_LINK_REGEX = /<a href="https:\/\/[\w.]+\/tags\/[\w]+" class="mention hashtag" rel="nofollow noopener noreferrer" target="_blank">#<span>[\w]+<\/span><\/a>/i;
+const HASHTAG_PARAGRAPH_REGEX = new RegExp(`^<p>(${HASHTAG_LINK_REGEX.source} ?)+</p>`, "i");
 
 const PROPS_THAT_CHANGE: KeysOfValueType<Toot, number>[] = [
     "favouritesCount",
@@ -290,6 +293,13 @@ export default class Toot implements TootObj {
         return this.mentions.some((mention) => mention.acct == MastoApi.instance.user.webfingerURI);
     }
 
+    // Return all but the last paragraph if that last paragraph is just hashtag links
+    contentNonTagsParagraphs(): string {
+        const paragraphs = htmlToParagraphs(this.content);
+        if (this.contentTagsParagraph()) paragraphs.pop(); // Remove the last paragraph if it's just hashtags
+        return paragraphs.join("\n");
+    }
+
     // Shortened string of content property stripped of HTML tags
     contentShortened(maxChars?: number): string {
         maxChars ||= MAX_CONTENT_PREVIEW_CHARS;
@@ -310,6 +320,16 @@ export default class Toot implements TootObj {
      // Return the toot's 'content' field stripped of HTML tags
     contentString(): string {
         return htmlToText(this.realToot().content || "");
+    }
+
+    // If the final <p> paragraph of the content is just hashtags, return it
+    contentTagsParagraph(): string | undefined {
+        const paragraphs = htmlToParagraphs(this.content || "");
+        const finalParagraph = paragraphs.slice(-1)[0];
+
+        if (HASHTAG_PARAGRAPH_REGEX.test(finalParagraph)) {
+            return finalParagraph;
+        }
     }
 
     // Return the toot's content + link description stripped of everything (links, mentions, tags, etc.)
@@ -445,6 +465,7 @@ export default class Toot implements TootObj {
 
         return this.resolvedToot;
     }
+
 
     // TODO: this maybe needs to take into consideration reblogsBy??
     tootedAt(): Date {